revert(chat): restore the existing scrolling behavior
