+++ packages/instantsearch-ui-components/src/lib/index.ts
@@ -1,6 +1,3 @@
 export * from './cx';
 export * from './stickToBottom';
 export * from './utils';
-
-/** @internal */
-export { createChatStickToBottom } from './chatStickToBottom';
