+++ packages/instantsearch-ui-components/src/lib/stickToBottom.ts
@@ -3,14 +3,32 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import {
-  createStickToBottomCore,
-  type StickToBottomCoreController,
-  type StickToBottomCoreRequest,
-  type StickToBottomCoreState,
-} from './stickToBottomCore';
-
 import type { Hooks } from '../types';
+
+export interface StickToBottomState {
+  scrollTop: number;
+  lastScrollTop?: number;
+  ignoreScrollToTop?: number;
+  targetScrollTop: number;
+  calculatedTargetScrollTop: number;
+  scrollDifference: number;
+  resizeDifference: number;
+
+  animation?: {
+    behavior: 'instant' | Required<SpringAnimation>;
+    ignoreEscapes: boolean;
+    promise: Promise<boolean>;
+  };
+  lastTick?: number;
+  velocity: number;
+  accumulated: number;
+
+  escapedFromLock: boolean;
+  isAtBottom: boolean;
+  isNearBottom: boolean;
+
+  resizeObserver?: ResizeObserver;
+}
 
 const DEFAULT_SPRING_ANIMATION = {
   /**
@@ -101,13 +119,24 @@ export type ScrollToBottom = (
 ) => Promise<boolean> | boolean;
 export type StopScroll = () => void;
 
-export interface StickToBottomState
-  extends Omit<StickToBottomCoreState, 'animation'> {
-  animation?: {
-    behavior: 'instant' | Required<SpringAnimation>;
-    ignoreEscapes: boolean;
-    promise: Promise<boolean>;
-  };
+const STICK_TO_BOTTOM_OFFSET_PX = 70;
+const SIXTY_FPS_INTERVAL_MS = 1000 / 60;
+const RETAIN_ANIMATION_DURATION_MS = 350;
+
+let mouseDown = false;
+
+if (typeof window !== 'undefined') {
+  window.document?.addEventListener('mousedown', () => {
+    mouseDown = true;
+  });
+
+  window.document?.addEventListener('mouseup', () => {
+    mouseDown = false;
+  });
+
+  window.document?.addEventListener('click', () => {
+    mouseDown = false;
+  });
 }
 
 export interface StickToBottomInstance {
@@ -126,62 +155,87 @@ type CreateStickToBottomParams = Pick<
   'useCallback' | 'useEffect' | 'useMemo' | 'useRef' | 'useState'
 >;
 
-const RETAIN_ANIMATION_DURATION_MS = 350;
-
-export function createStickToBottom(hooks: CreateStickToBottomParams) {
-  const useStickToBottomCore = createStickToBottomCore(hooks);
-
+export function createStickToBottom({
+  useCallback,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+}: CreateStickToBottomParams) {
   return function useStickToBottom(
     options: StickToBottomOptions = {}
   ): StickToBottomInstance {
-    const optionsRef = hooks.useRef<StickToBottomOptions>(null!);
+    const [escapedFromLock, updateEscapedFromLock] = useState(false);
+    const [isAtBottom, updateIsAtBottom] = useState(options.initial !== false);
+    const [isNearBottom, setIsNearBottom] = useState(false);
+
+    const optionsRef = useRef<StickToBottomOptions>(null!);
     optionsRef.current = options;
 
-    const controller = hooks.useMemo<
-      StickToBottomCoreController<ScrollToBottomOptions>
-    >(() => {
+    // Create refs early so they can be used in other hooks
+    const scrollRef = useRef<HTMLDivElement | null>(null);
+    const contentRef = useRef<HTMLDivElement | null>(null);
+
+    const isSelecting = useCallback(() => {
+      if (!mouseDown) {
+        return false;
+      }
+
+      if (typeof window === 'undefined') {
+        return false;
+      }
+
+      const selection = window.getSelection();
+      if (!selection || !selection.rangeCount) {
+        return false;
+      }
+
+      const range = selection.getRangeAt(0);
+      return (
+        range.commonAncestorContainer.contains(scrollRef.current) ||
+        scrollRef.current?.contains(range.commonAncestorContainer)
+      );
+    }, []);
+
+    // biome-ignore lint/correctness/useExhaustiveDependencies: state is intentionally stable
+    const state = useMemo<StickToBottomState>(() => {
       let lastCalculation:
         | { targetScrollTop: number; calculatedScrollTop: number }
         | undefined;
 
-      const createRequest = (
-        scrollOptions: ScrollToBottomOptions = {}
-      ): StickToBottomCoreRequest => {
-        if (typeof scrollOptions === 'string') {
-          scrollOptions = { animation: scrollOptions };
-        }
-
-        const waitElapsed = Date.now() + (Number(scrollOptions.wait) || 0);
-        const request: StickToBottomCoreRequest = {
-          behavior: mergeAnimations(
-            optionsRef.current,
-            scrollOptions.animation
-          ),
-          waitElapsed,
-          waitForCurrent: scrollOptions.wait === true,
-          ignoreEscapes: scrollOptions.ignoreEscapes ?? false,
-          preserveScrollPosition: scrollOptions.preserveScrollPosition ?? false,
-        };
-
-        if (scrollOptions.duration instanceof Promise) {
-          scrollOptions.duration.then(
-            () => {
-              request.durationElapsed = Date.now();
-            },
-            () => {
-              request.durationElapsed = Date.now();
-            }
-          );
-        } else {
-          request.durationElapsed = waitElapsed + (scrollOptions.duration ?? 0);
-        }
-
-        return request;
-      };
-
       return {
-        initialIsAtBottom: optionsRef.current.initial !== false,
-        calculateTargetScrollTop(targetScrollTop, context) {
+        escapedFromLock,
+        isAtBottom,
+        resizeDifference: 0,
+        accumulated: 0,
+        velocity: 0,
+
+        get scrollTop() {
+          return scrollRef.current?.scrollTop ?? 0;
+        },
+        set scrollTop(scrollTop: number) {
+          if (scrollRef.current) {
+            scrollRef.current.scrollTop = scrollTop;
+            state.ignoreScrollToTop = scrollRef.current.scrollTop;
+          }
+        },
+
+        get targetScrollTop() {
+          if (!scrollRef.current || !contentRef.current) {
+            return 0;
+          }
+
+          return (
+            scrollRef.current.scrollHeight - 1 - scrollRef.current.clientHeight
+          );
+        },
+        get calculatedTargetScrollTop() {
+          if (!scrollRef.current || !contentRef.current) {
+            return 0;
+          }
+
+          const { targetScrollTop } = this;
+
           if (!optionsRef.current.targetScrollTop) {
             return targetScrollTop;
           }
@@ -192,7 +246,10 @@ export function createStickToBottom(hooks: CreateStickToBottomParams) {
 
           const calculatedScrollTop = Math.max(
             Math.min(
-              optionsRef.current.targetScrollTop(targetScrollTop, context),
+              optionsRef.current.targetScrollTop(targetScrollTop, {
+                scrollElement: scrollRef.current,
+                contentElement: contentRef.current,
+              }),
               targetScrollTop
             ),
             0
@@ -206,37 +263,391 @@ export function createStickToBottom(hooks: CreateStickToBottomParams) {
 
           return calculatedScrollTop;
         },
-        createScrollRequest: createRequest,
-        createResizeRequest(initial) {
-          const behavior = mergeAnimations(
-            optionsRef.current,
-            initial ? optionsRef.current.initial : optionsRef.current.resize
-          );
 
-          return createRequest({
-            animation: behavior,
-            wait: true,
-            preserveScrollPosition: true,
-            duration:
-              behavior === 'instant' ? undefined : RETAIN_ANIMATION_DURATION_MS,
-          });
+        get scrollDifference() {
+          return this.calculatedTargetScrollTop - this.scrollTop;
         },
-        createContinuationRequest(ignoreEscapes, duration) {
-          return createRequest({
-            animation: mergeAnimations(
-              optionsRef.current,
-              optionsRef.current.resize
-            ),
-            ignoreEscapes,
-            duration,
-          });
+
+        get isNearBottom() {
+          return this.scrollDifference <= STICK_TO_BOTTOM_OFFSET_PX;
         },
       };
-      // Keep the controller stable while callbacks read current options through optionsRef.
       // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
-    return useStickToBottomCore(controller) as StickToBottomInstance;
+    const setIsAtBottom = useCallback(
+      (value: boolean) => {
+        state.isAtBottom = value;
+        updateIsAtBottom(value);
+      },
+      [state]
+    );
+
+    const setEscapedFromLock = useCallback(
+      (value: boolean) => {
+        state.escapedFromLock = value;
+        updateEscapedFromLock(value);
+      },
+      [state]
+    );
+
+    const scrollToBottom = useCallback<ScrollToBottom>(
+      (scrollOptions = {}) => {
+        if (typeof scrollOptions === 'string') {
+          scrollOptions = { animation: scrollOptions };
+        }
+
+        if (!scrollOptions.preserveScrollPosition) {
+          setIsAtBottom(true);
+        }
+
+        const waitElapsed = Date.now() + (Number(scrollOptions.wait) || 0);
+        const behavior = mergeAnimations(
+          optionsRef.current,
+          scrollOptions.animation
+        );
+        const { ignoreEscapes = false } = scrollOptions;
+
+        let durationElapsed: number;
+        let startTarget = state.calculatedTargetScrollTop;
+
+        if (scrollOptions.duration instanceof Promise) {
+          scrollOptions.duration.then(
+            () => {
+              durationElapsed = Date.now();
+            },
+            () => {
+              durationElapsed = Date.now();
+            }
+          );
+        } else {
+          durationElapsed = waitElapsed + (scrollOptions.duration ?? 0);
+        }
+
+        const next = (): Promise<boolean> => {
+          const promise = new Promise(requestAnimationFrame).then(() => {
+            if (!state.isAtBottom) {
+              state.animation = undefined;
+
+              return false;
+            }
+
+            const { scrollTop } = state;
+            const tick = performance.now();
+            const tickDelta =
+              (tick - (state.lastTick ?? tick)) / SIXTY_FPS_INTERVAL_MS;
+            state.animation ||= { behavior, promise, ignoreEscapes };
+
+            if (state.animation.behavior === behavior) {
+              state.lastTick = tick;
+            }
+
+            if (isSelecting()) {
+              return next();
+            }
+
+            if (waitElapsed > Date.now()) {
+              return next();
+            }
+
+            if (
+              scrollTop < Math.min(startTarget, state.calculatedTargetScrollTop)
+            ) {
+              if (state.animation?.behavior === behavior) {
+                if (behavior === 'instant') {
+                  state.scrollTop = state.calculatedTargetScrollTop;
+                  return next();
+                }
+
+                state.velocity =
+                  (behavior.damping * state.velocity +
+                    behavior.stiffness * state.scrollDifference) /
+                  behavior.mass;
+                state.accumulated += state.velocity * tickDelta;
+                state.scrollTop += state.accumulated;
+
+                if (state.scrollTop !== scrollTop) {
+                  state.accumulated = 0;
+                }
+              }
+
+              return next();
+            }
+
+            if (durationElapsed > Date.now()) {
+              startTarget = state.calculatedTargetScrollTop;
+
+              return next();
+            }
+
+            state.animation = undefined;
+
+            /**
+             * If we're still below the target, then queue
+             * up another scroll to the bottom with the last
+             * requested animatino.
+             */
+            if (state.scrollTop < state.calculatedTargetScrollTop) {
+              return scrollToBottom({
+                animation: mergeAnimations(
+                  optionsRef.current,
+                  optionsRef.current.resize
+                ),
+                ignoreEscapes,
+                duration:
+                  Math.max(0, durationElapsed - Date.now()) || undefined,
+              });
+            }
+
+            return state.isAtBottom;
+          });
+
+          return promise.then((result) => {
+            requestAnimationFrame(() => {
+              if (!state.animation) {
+                state.lastTick = undefined;
+                state.velocity = 0;
+              }
+            });
+
+            return result;
+          });
+        };
+
+        if (scrollOptions.wait !== true) {
+          state.animation = undefined;
+        }
+
+        if (state.animation?.behavior === behavior) {
+          return state.animation.promise;
+        }
+
+        return next();
+      },
+      [setIsAtBottom, isSelecting, state]
+    );
+
+    const stopScroll = useCallback((): void => {
+      setEscapedFromLock(true);
+      setIsAtBottom(false);
+    }, [setEscapedFromLock, setIsAtBottom]);
+
+    const handleScroll = useCallback(
+      ({ target }: Event) => {
+        if (target !== scrollRef.current) {
+          return;
+        }
+
+        const { scrollTop, ignoreScrollToTop } = state;
+        let { lastScrollTop = scrollTop } = state;
+
+        state.lastScrollTop = scrollTop;
+        state.ignoreScrollToTop = undefined;
+
+        if (ignoreScrollToTop && ignoreScrollToTop > scrollTop) {
+          /**
+           * When the user scrolls up while the animation plays, the `scrollTop` may
+           * not come in separate events; if this happens, to make sure `isScrollingUp`
+           * is correct, set the lastScrollTop to the ignored event.
+           */
+          lastScrollTop = ignoreScrollToTop;
+        }
+
+        setIsNearBottom(state.isNearBottom);
+
+        /**
+         * Scroll events may come before a ResizeObserver event,
+         * so in order to ignore resize events correctly we use a
+         * timeout.
+         *
+         * @see https://github.com/WICG/resize-observer/issues/25#issuecomment-248757228
+         */
+        setTimeout(() => {
+          /**
+           * When theres a resize difference ignore the resize event.
+           */
+          if (state.resizeDifference || scrollTop === ignoreScrollToTop) {
+            return;
+          }
+
+          if (isSelecting()) {
+            setEscapedFromLock(true);
+            setIsAtBottom(false);
+            return;
+          }
+
+          const isScrollingDown = scrollTop > lastScrollTop;
+          const isScrollingUp = scrollTop < lastScrollTop;
+
+          if (state.animation?.ignoreEscapes) {
+            state.scrollTop = lastScrollTop;
+            return;
+          }
+
+          if (isScrollingUp) {
+            setEscapedFromLock(true);
+            setIsAtBottom(false);
+          }
+
+          if (isScrollingDown) {
+            setEscapedFromLock(false);
+          }
+
+          if (!state.escapedFromLock && state.isNearBottom) {
+            setIsAtBottom(true);
+          }
+        }, 1);
+      },
+      [setEscapedFromLock, setIsAtBottom, isSelecting, state]
+    );
+
+    const handleWheel = useCallback(
+      ({ target, deltaY }: WheelEvent) => {
+        let element = target as HTMLElement;
+
+        while (
+          !['scroll', 'auto'].includes(getComputedStyle(element).overflow)
+        ) {
+          if (!element.parentElement) {
+            return;
+          }
+
+          element = element.parentElement;
+        }
+
+        /**
+         * The browser may cancel the scrolling from the mouse wheel
+         * if we update it from the animation in meantime.
+         * To prevent this, always escape when the wheel is scrolled up.
+         */
+        if (
+          element === scrollRef.current &&
+          deltaY < 0 &&
+          scrollRef.current.scrollHeight > scrollRef.current.clientHeight &&
+          !state.animation?.ignoreEscapes
+        ) {
+          setEscapedFromLock(true);
+          setIsAtBottom(false);
+        }
+      },
+      [setEscapedFromLock, setIsAtBottom, state]
+    );
+
+    // Attach scroll and wheel event listeners
+    useEffect(() => {
+      const scroll = scrollRef.current;
+      if (!scroll) {
+        return undefined;
+      }
+
+      scroll.addEventListener('scroll', handleScroll, { passive: true });
+      scroll.addEventListener('wheel', handleWheel, { passive: true });
+
+      return () => {
+        scroll.removeEventListener('scroll', handleScroll);
+        scroll.removeEventListener('wheel', handleWheel);
+      };
+    }, [handleScroll, handleWheel]);
+
+    // Attach ResizeObserver to content element
+    useEffect(() => {
+      const content = contentRef.current;
+      if (!content) {
+        return undefined;
+      }
+
+      let previousHeight: number | undefined;
+
+      const resizeObserver = new ResizeObserver(([entry]) => {
+        const { height } = entry.contentRect;
+        const difference = height - (previousHeight ?? height);
+
+        state.resizeDifference = difference;
+
+        /**
+         * Sometimes the browser can overscroll past the target,
+         * so check for this and adjust appropriately.
+         */
+        if (state.scrollTop > state.targetScrollTop) {
+          state.scrollTop = state.targetScrollTop;
+        }
+
+        setIsNearBottom(state.isNearBottom);
+
+        if (difference >= 0) {
+          /**
+           * If it's a positive resize, scroll to the bottom when
+           * we're already at the bottom.
+           */
+          const animation = mergeAnimations(
+            optionsRef.current,
+            previousHeight
+              ? optionsRef.current.resize
+              : optionsRef.current.initial
+          );
+
+          scrollToBottom({
+            animation,
+            wait: true,
+            preserveScrollPosition: true,
+            duration:
+              animation === 'instant'
+                ? undefined
+                : RETAIN_ANIMATION_DURATION_MS,
+          });
+        } else if (state.isNearBottom) {
+          /**
+           * Else if it's a negative resize, check if we're near the bottom
+           * if we are want to un-escape from the lock, because the resize
+           * could have caused the container to be at the bottom.
+           */
+          setEscapedFromLock(false);
+          setIsAtBottom(true);
+        }
+
+        previousHeight = height;
+
+        /**
+         * Reset the resize difference after the scroll event
+         * has fired. Requires a rAF to wait for the scroll event,
+         * and a setTimeout to wait for the other timeout we have in
+         * resizeObserver in case the scroll event happens after the
+         * resize event.
+         */
+        requestAnimationFrame(() => {
+          setTimeout(() => {
+            if (state.resizeDifference === difference) {
+              state.resizeDifference = 0;
+            }
+          }, 1);
+        });
+      });
+
+      resizeObserver.observe(content);
+      state.resizeObserver = resizeObserver;
+
+      return () => {
+        resizeObserver.disconnect();
+        state.resizeObserver = undefined;
+      };
+    }, [
+      state,
+      setIsNearBottom,
+      setEscapedFromLock,
+      setIsAtBottom,
+      scrollToBottom,
+    ]);
+
+    return {
+      contentRef,
+      scrollRef,
+      scrollToBottom,
+      stopScroll,
+      isAtBottom: isAtBottom || isNearBottom,
+      isNearBottom,
+      escapedFromLock,
+      state,
+    };
   };
 }
