+++ packages/instantsearch.js/src/widgets/chat/chat.tsx
@@ -15,7 +15,7 @@ import {
   DisplayResultsToolType,
 } from '../../lib/chat';
 import { prepareTemplateProps } from '../../lib/templating';
-import { useChatStickToBottom } from '../../lib/useChatStickToBottom';
+import { useStickToBottom } from '../../lib/useStickToBottom';
 import {
   getContainerNode,
   createDocumentationMessageGenerator,
@@ -225,7 +225,10 @@ function ChatWrapper({
   state,
 }: ChatWrapperProps) {
   const { scrollRef, contentRef, scrollToBottom, isAtBottom } =
-    useChatStickToBottom();
+    useStickToBottom({
+      initial: 'smooth',
+      resize: 'smooth',
+    });
 
   // Keep the conversation pinned to the bottom while streaming. The stick-to-
   // bottom ResizeObserver only reacts to content *height* changes, but tool
