--- packages/react-instantsearch/src/lib/useChatStickToBottom.ts
@@ -1,15 +1,0 @@
-/* !---------------------------------------------------------------------------------------------
- *  Copyright (c) StackBlitz. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { createChatStickToBottom } from 'instantsearch-ui-components';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-
-export const useChatStickToBottom = createChatStickToBottom({
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-});
